Separate commands into their own file.
