+++ gits/utils/__init__.py
@@ -1,0 +1,2 @@
+# gits/__init__.py
+# Marks this directory as a Python package.
